Move remaining workflow client modules to corresponding subdirectory
