+++ client/galaxy/scripts/components/Workflow/Editor/modules/canvas.js
