+++ client/galaxy/scripts/components/Workflow/Editor/modules/connector.js
