+++ client/galaxy/scripts/components/Workflow/Editor/modules/outputs.js
