+++ client/galaxy/scripts/components/Workflow/Editor/modules/services.js
