+++ client/galaxy/scripts/components/Workflow/Editor/modules/terminals.js
